chore: remove script condition helper from shared
