+++ packages/shared/src/index.ts
@@ -3,14 +3,12 @@ import deepmerge from '../compiled/deepmerge/index.js';
 import color from '../compiled/picocolors/index.js';
 import RspackChain from '../compiled/rspack-chain/index.js';
 import type { CacheGroups } from './types';
-import type { NormalizedEnvironmentConfig, RsbuildContext } from './types';
 
 export * from './types';
 
 // RegExp
 export const JS_REGEX: RegExp = /\.(?:js|mjs|cjs|jsx)$/;
 export const SCRIPT_REGEX: RegExp = /\.(?:js|jsx|mjs|cjs|ts|tsx|mts|cts)$/;
-export const TS_AND_JSX_REGEX: RegExp = /\.(?:ts|tsx|jsx|mts|cts)$/;
 export const NODE_MODULES_REGEX: RegExp = /[\\/]node_modules[\\/]/;
 
 enum ESVersion {
@@ -296,45 +294,3 @@ export const CHAIN_ID = {
 } as const;
 
 export type ChainIdentifier = typeof CHAIN_ID;
-
-export function applyScriptCondition({
-  rule,
-  chain,
-  config,
-  context,
-  includes,
-  excludes,
-}: {
-  rule: RspackChain.Rule;
-  chain: RspackChain;
-  config: NormalizedEnvironmentConfig;
-  context: RsbuildContext;
-  includes: (string | RegExp)[];
-  excludes: (string | RegExp)[];
-}): void {
-  // compile all folders in app directory, exclude node_modules
-  // which can be removed next version of rspack
-  rule.include.add({
-    and: [context.rootPath, { not: NODE_MODULES_REGEX }],
-  });
-
-  // Always compile TS and JSX files.
-  // Otherwise, it will lead to compilation errors and incorrect output.
-  rule.include.add(TS_AND_JSX_REGEX);
-
-  // The Rsbuild runtime code is es2017 by default,
-  // transform the runtime code if user target < es2017
-  const target = castArray(chain.get('target'));
-  const legacyTarget = ['es5', 'es6', 'es2015', 'es2016'];
-  if (legacyTarget.some((item) => target.includes(item))) {
-    rule.include.add(/[\\/]@rsbuild[\\/]core[\\/]dist[\\/]/);
-  }
-
-  for (const condition of [...includes, ...(config.source.include || [])]) {
-    rule.include.add(condition);
-  }
-
-  for (const condition of [...excludes, ...(config.source.exclude || [])]) {
-    rule.exclude.add(condition);
-  }
-}
